+++ www/commit-hash.js
@@ -1,1 +1,1 @@
-window.__COMMIT__ = '9f445c6ff6c51f53e63970991b193ee16e89d668';
+window.__COMMIT__ = '39ab701ef5c2059242d8a5d9e93737413df88c01';